% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distcrete.R
\name{distcrete}
\alias{distcrete}
\title{Discretise a distribution}
\usage{
distcrete(name, interval, ..., w = 0.5, anchor = 0)
}
\arguments{
\item{name}{The name of a distribution function (e.g.,
\code{norm}, \code{gamma}).  The distribution must have a cdf
function (e.g., \code{pnorm}) and a quantile function (e.g.,
\code{qnorm}) defined.}

\item{interval}{The interval to discretise the interval onto.}

\item{...}{Parameters to \code{cdf}.  Can be matched positionally
or by name.}

\item{w}{How to weight the endpoints; must be between 0 and 1.
If 0.5 then integration happens centred around the interval, if
0 floor, if 1 then ceiling.}

\item{anchor}{Any location that is a valid \code{x}}
}
\description{
Discretise a distribution.
}
\details{
{
\itemize{
\item
\eqn{x_{c}}: a random variable with a known continous distribution, with
\eqn{f_{c}} its p.d.f and \eqn{F_{c}} the corresponding c.d.f.
\item
\eqn{x_{d}}: a random variable following a discrete distribution derived
from \eqn{f_{c}}
}

The p.m.f. of \eqn{x_{d}} is defined for any discrete value \eqn{X} by: 
 \deqn{
   p(x_{d} = X) = p(x_{c} \in \left[X - (wi/2) ; X + (wi/2)\left[) = F_c(X + wi/2) - F_c(X - wi/2)
  } where \eqn{i} is the width of the discretising interval, and \#'(w} is
an offset describing how continuous values are mapped onto a discrete
scale. For instance, with \eqn{i = 1} and \eqn{w = 0.5}, any values on
\eqn{\left[2.5 ; 3.5\left[} are mapped to 2; for \eqn{i = 1} and \eqn{w = 0\
), any values on \eqn{\left[1 ; 2\left[} is mapped to \eqn{1}.
}
}
\examples{
set.seed(415)
d0 <- distcrete::distcrete("gamma", 1, shape = 3, w = 0)
d0$d(1:10)
d0$p(c(.1,.5))
d0$q(c(.1,.5))
d0$r(10)

}
\author{
Rich FitzJohn
}
